+++ EulerLib/Sequences/DeficientNumbers.cs
@@ -6,7 +6,7 @@ namespace EulerLib.Sequences
     public class DeficientNumbers
         : NaiveSequenceGeneratorBase
     {
-        public override bool IntegerIsMemberOfSequence(int n)
+        public override bool IntegerIsMemberOfSequence(long n)
         {
             return n.IsDeficient();
         }
