+++ EulerLib/Sequences/PerfectNumbers.cs
@@ -5,7 +5,7 @@ namespace EulerLib.Sequences
 {
     public class PerfectNumbers : NaiveSequenceGeneratorBase
     {
-        public override bool IntegerIsMemberOfSequence(int n)
+        public override bool IntegerIsMemberOfSequence(long n)
         {
             return n.IsPerfect();
         }
